Add script to est. spectral phasor using LeastSq
In this commit, I have added a couple of scripts which compare the
performance of the complex phasor associated with a particular frequency estimated using FFT and using least squares method. Also, in the sai_radar_model_testbench.py script, there was a bug in obtaining the signal amplitudes from the noise variance and snr. To obtain the weights/ amplitudes, there should be a square root. This has been fixed in this commit.

Signed-off-by: Sai Gunaranjan <saiguna@steradiansemi.com>