Compute SNR as a bin SNR unlike conven. snr defn.
In this commit, I have modified the way the snr and hence signal weights
of a signal in noise are computed. The conventional definition of SNR is
the ratio of signal power to the noise power. However, when dealing with
spectrum, we typically define the bin snr to be the SNR of the object.
So in a way, it is the ratio of the signal power to the average noise
power per (FFT) bin. Taking this forward, I have mode the following
modifications to the snr based scripts.
Most importantly, for the 'radar_model_testbench.py' script, I am
computing the signal weights as follows
Previously, the Noise power was arbitrarily fixed to some value. Now, the Noise power is modelled along the lines of the receiver RF signal chain which includes the thermal noise of -174 dBm/Hz, receiver gain, noise figure, IF bandwidth.The sum of the above qunatities(in the dB scale) gives the total noise power (dBm) in the IF bandwidth. Hence the noise power bin is the ratio of the total noise power to the inherent number of signal samples(or number of FFT bins). The signal weights are then computed as 10 pow((noiseFloorPerBin_dB + SNR)/10).
I have made similar modifications to the other scripts which use snr
definitions.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>