Converge OMP based on Noise Floor per bin
In this commit, I have added a Noise floor based convergence for all the
OMP related scripts. Even now, it is not converging to the noise floor
after running for number of iterations = number of objects. This has to
be debugged further. However, now atleast it is based on some logic
unlike previously when an arbitrary small value was set as a convergence
number. A word of caution though. I have not applied the same
convergence criterion to the Matching Pursuit(MP).

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>