LeastSquares based CoeffEst as f(freq separation)
In this commit, we analyse the performance of the Least Squares based
method in estimating the complex coefficients/phasors of two known
frequencies as the separation between them is varied and also the input
SNR is varied. We compare the true/ground truth coefficients with those
estimated from least squares.

Signed-off-by: Sai Gunaranjan Pelluri <saiguna@steradiansemi.com>