Model impact of phase jump in a linear phase signal on the spectrum of the signal
In this commit, I have modelled the impact of a phase jump on the
spectrum of the signal. For this, I consider a linear phase signal and
introduce a small jump in the phase of the signal. The I plot the
magnitude spectru of the signal. The phase jump appears as a small
kink/sidelobe on the spectrum and thus reducing the dynamic range of the
signal spectrum. Turns out if the phase jump is equal to pi radians,
then the single peak splits into two false peaks. Also, changing the
sign of the phase jump(from positive to negative) aslo changes the
location of the kink from one side to the other. I have illustrated
these phenomenon in the script with appropriate plots.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>