Add a script to model interaction of 2 frequencies sampled at 2 different rates
In this commit, I have added a script to model the interaction between
two frequencies/Velocities(in RADAR terminology) in the same range bin
and sampled at two different rates. When sampled at a rate say Fs1, the
signal at baseband could be coming from any of the Hypothesis m. The
true frequency f = fb + M*Fs1 is received at baseband as fb. From this
alone, we cannot say what the true frequency could be. When also sampled
at Fs2 rate, the frequency f occurs at the baseband frequency fb_ =
modulo(f,Fs2). Hence if we have only a single frequency component, we
can identify the true freqeuncy just by sampling at 2 different rates
and observing the baseband frequencies at each of the 2 rates. However,
when there are multiple frequencies, then the baseband frequency of one
signal (which we refer to as aggressor) biases/falls on the hypothesised
freqeuncy of the other signal component (referred to as victim) when
sampled at the Fs2 rate. In the context of RADARs, when a range bin has
2 Dopplers, then 1 acts as an aggressor and biases the prediction of the
Doppler integer Hypothesis of the other Doppler. This script essentially
models all the aggressor velocities for a given victim frequency. The
steps to model are mentioned in the script as a docstring.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>