Model the periodicity in Phase Shifter due to the DNL
In this commit, I have modeled how the phase shifter error(from the
ground truth)for each unique phase shifter manifests as periodic repetetions when the phase is swept over several ramps. This periodic component for each of the phase shifter manifests as spurs in the Doppler spectrum. These spurs get aggravated when the phase delta(in degrees) programmed per ramp divides 360 degress exactly. In such case, the same phase shifter code ia applied periodically at every 360/phase_delta ramp number. For example, if the phase delta per ramp is chosen as 30 degrees, then 360/30 = 12 i.e. the starting ramp and the 12th ramp will have the same phase shifter applied. Similarly, the 2nd ramp and the 13th ramp have the same phase shifter applied and so on. Now since each phase shifter has its own "signature"(error in terms of DNL wrt ground truth), there will be a periodic pattern every 12th ramp. To break this periodicity, I have chosen a phase step which doesn't divide 360 degrees. For example, if the phase step per ramp is 29 degress instead of 30 degress, then at the 13th ramp, the phase is 29*13 = 377 = 17 degress. But this is not equal to 0 degrees i.e the phase shifter for the 13th ramp is not the same as the phase shifter for the 1st ramp. Similarly, the second ramp has a phase of 29 degress while the 14th ramp has a phase of 29*14= 406 = 46 degress which is not the same as 29 degress. Hence the phase shifter for the 14th ramp is not the same as the phase shifter for the 2nd ramp. Hence just by making the phase change per ramp not a divisor of 360 degress, we remove the periodicity.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>