Model simultaneous transmission from 4Txs each with different phase step per ramp
In this commit, I have modelled the simultaneous transmission from 4 Txs each with its own phase step across ramps. Hence the Doppler spectrum will contain 4 peaks with the object Doppler modulated to the location of Phase step corresponding to each Tx. Now, since each Tx has its own ramp phase step, the periodicity of the Phase shifter quantization also differs for each Tx. For example in this case where the phase step per Tx is chosen as 30 degrees, the 0th Tx has no phase sweep across ramps and hence the contribution of the 0th Tx to the Doppler spectrum is just a DC. The 1st Tx has a phase step of 30 degress per ramp and hence the Phase shifter repeats every 12th ramp. Now, we had seen in the previous commit that the periodicity can be broken by adding a small quadratic phase term which is controlled by the weight alpha. We had seen that an alhoa of 0.0001 was a good operating point (large enough to break the periodicity but small enough to not cause a spreading of the peak). But when it comes to the 2nd and 3rd Tx has have a phase step of 2*30 and 3*30, the same alpha of 0.0001 may be be the right choice. Hence the alpha term should not be a constant but rather a function of the phase step per ramp and the Tx. In other words, it should be a function of p*deltaPhi, where p is the ruiing index for the Tx and deltaPhi is the phase step per Tx per ramp. In the current commit, I have put a constant alpha for each of the Tx. In the next commit, I will attempt to optimize and get an alpha that should be applied to each Tx.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>