Observed 3 dB raise in noiseFloor with each new Tx added.
I have observed a 3 dB raise in noise floor with an addition of a new Tx to the simultaneous Code based Tx transmission. In other words, if we have 2 Txs transmitting perviously and I add a 3rd Tx to transmit imultaneously with its own phase sweep, the noise floor raises by 3 dB. Similarly going from 3 Txs to 4 Txs also increases the noise floor by another 3 dB. Not sure why this is happening. Need to understand this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>