Fix bug in the Doppler bin computation.
Previosuly, I was taking modulo Fs/2 of the velocity and post this, I
was converting the velocity to bipolar. This is wrong. In this commit, I have fixed this bug. Now, I take the modulo of Fs of the velocity and the compute the bin as is without making it bi-polar. Also, in this commit, I have put all the constants like system parameters at the beginning of the code. This make the code a lot cleaner.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>