Illustrate how range bin movement across chirps induces phase jumps on the signal bin
In this commit, we look at the phase of the signal range bin and how it varies as the range bin moves across chips for a high speed target. Previously, we had modelled the range bin migration term into the RADAR equation to sumulate an actual scenario when there is a high speed target. In this commit, we track the signal range bin across chips and then try to analyze the phase behaviour on the signal bin (as the bin moves across chirps). We observe that at those chirps where the signal peak changes by 1 bin, a phase jump is introduced. For example, let us say for the first 10 chirps, the peak in the range spectrum gets quantized to the same range bin but at the 11th chirp, the range bin peak shifts by 1 bin. Now, if we sample the peak bin for every chirp and plot the unwrapped phase on the signal bin across chirps, we observe a phase jump at precisely those chirps where the signal peak has moved to the adjacent range bin. So basically, what this implies is that, as the signal migrates from one range bin to another, a phase jump is introduced. When we run this script, we can see the plots. We need to understand why this phase jump occurs and what is the magnitude of this phase jump. As a co-incidence, it turns out that the phase jump is very close to pi radians! And so, when I applied a phase correction of pi radians to the signal phase everytime there is bin migration and then replotted the signal phase, it was alomst prefectly linear. What a co-incidence! But we need to understand why the phase jump is pi radians. Is it exactly pi radians or someting clode to pi radians? To answer these questions, I set out trying to derive the closed form expression for the phase jump when a bin migration occurs.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>