Model Frequency bin migration and analyse the phase jump introduced due to it
In the previous commit, I had used the radar base-band equation to model the target range, Doppler, range bin migration term and from this model, I had plotted the signal phase across chirps by sampling the peak range bin across chirps. So the phase across chirps had 2 components. 1) The inherent Doppler phase 2) The residual phase between adjacent bins as the signal peaks traverses from 1 range bin to the other. So this script was analysing the phase for a radar model based target. However, the phase jump due to bin migration can be modelled standalone and has no dependence on RADAR eqaution per se. We can simply simulate several instances of a signal but with the frequency increasing linearly for each instance of the signal such that for the middle instance of the signal, the peak would have moved to the adjacent bin. So, the first signal will have a freqeuncy of f. The second signal will have a frequency of f+delta, the third signal will have a freqeuncy of f+2delta and so on. This model is a simplistic model and unlike the radar based model is devoid of Doppler term. The phase across different instances of the signal contains only one component i.e. the residual phase between adjacent bins as the signal peaks traverses from 1 range bin to the other. So in this commit, I have added a new script named 'frequencyBinMigrationInducedPhaseJumps_standAloneSimulation' which analyses the phase across chirps/instances for a simplistic sinusoidal signal whose frequency keeps slighly shifting from one instance to the other. From this simulation, we make an important observation that the phase delta between the true bin and the leakage bin is always constant no matter which instance of the signal we take. In other words, if we have a signal frequency peak not falling exaclty on a bin, the difference in phase between the true bin(closest to the true frequency) and the adjacent leakage bin will always be a constant ~pi radians. In the next commit, I will derive the closed form expression for the phase delta and its exact value.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>