Model and derive the phase jump introduced due to frequency bin migration
The phase jump introduced due to range bin migration is essentally the same problem as finding the phase delta between the signal bin and its leakage/adjacent bin. When a frequency component moves from one frequency bin to the other, there is a phase change introduced. Here, we analyse the residual phase on bins $k_{0}$ and $k_{0}+1$ as the peak moves from bin $k_{0}$ to bin $k_{0}+1$. Let us assume we have a signal whose frequency bin is given by $k_{0}+\alpha$ and $\alpha$ varies from 0 to 1. When $\alpha = 0$, the peak occurs exactly at bin $k_{0}$ and when $\alpha = 1$, the peak occurs exactly at bin $k_{0}+1$. Our objective is to study the phase at bins $k_{0}$ and $k_{0}+1$ as $\alpha$ varies from 0 to 1. In other words, we want to understand the phase behaviour of the leakage bins as the peak moves from $k_{0}$ to $k_{0}+1$ through $\alpha$. The theoreical derivation for the closed form expression for the phase delta between the signal freqeuncy bin and the leakage bin which is essentially the same as the phase jump due to frequency bin migration is mentioned as a link in the script. I will also attach the pdf of the derivation with this commit.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>