Plot and analyse response of Inverse Vandermonde Matrix on the bipolar frequency axis scale from -pi to pi
In this commit, I analyse the frequency response/zeros of the Inverse Vandermonde rows on a bipolar frequency grid i.e from -pi to +pi. In the previous commit, the analysis was being done from 0 to 2pi. I have also added some useful comments to explain some important lines of code. The following are the other changes made in this commit.
1. Added a comment explaining why we take negative of the phases for the signal frequencies
2. Applying an FFTshift on the FFT of the Inverse of the vandermonde matrix. This is so we cna work on the bipolar freqeuncy axis from -pi to +pi
3. Changed the plotting color and style.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>