Synthesize a pseudo inverse based response by manipulating the zeros of a window function
In this commit, I have developed a script which synthesizes a pseudo inverse like response(of Vandermonde matrix) by manipulating the zeros of a pre-defined simple function. In this case, we have used a kaiser window to start with. We treat the window function as a system finite impulse response(FIR) and find the zeros of the window function. We then move the system zeros to the position of the interferer frequency(digital) locations. In fact, the zeros have to be pushed to the negative of the interferer locations. This is because, we need to synthesize a vandermonde pseudo inverse like response which has an inherent conjugation and hence the frequency response is also flipped. So we need to place the zeros at the negative of the interferer locations. For this, we need to identify the system zeros closest to the interferer tones and this closeness measure has to be done modulo 2pi. For example, lets say 2 of the system zeros to begin with are located at 0 deg and 330 degrees. Now say there is an interferer at 358 degrees. Then, if we were to just take the smallest magnitude difference of the phases, then the interferer would seem closer to 330 degress than 0 degrees since (358-330 < 358-0). But in reality, 358 degrees or -2 degrees is closer to 0. So in order to take care of such cases, we use a modulo difference to find the closest zeros to the interferer tones(actually negative of the interferer tones) and replace the closest system zeros with the negative of the interferer. Once the zeros are manipulated, we again reconstruct the impulse response using the zpk2tf function from the scipy.signal library. In this script, I have taken a simple case of 2 tones. Hence we get two impulse responses. For the first impulse response, the second tone is the interferer and it is nulled out with the nearest zero and for the second impulse response, the first tone becomes the interferer. I have plotted the frequency response of the synthesised impulse responses and overlaid them with the vandermonde pseudo inverse response. The location of the nulls/zeros is matching expectation. One thing I'm circumspect about is the synthesised impulse response correlation with the observaion to get the complex coefficients. The absolute values will not match the least squares based solution but will the structure of the phase across MIMO channels be maintained? I will need to simulate to test this aspect.

In the next commit, I will replace the zpk2tf function with my own function which does an iterative convolution treating the overall reponse as a series of cascaded FIR filters each with a single zero.

Note: An FIR system with N taps, will have N-1 zeros.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>