Replace scipy.signal.zpk2tf with my own tf generating function
In this commit, I have written my own function which takes in the zeros of an FIR system as input and generates the equivalent impulse response of the system. The length of the generated impulse response is the number of zeros of the sytem + 1. This function can be used only for FIR systems which have only zeros and no poles. Here's how this function works.
Let us assume the system has 13 zeros. Each zero of the system is treated as a simple 2 tap FIR filter(N zeros implies a N+1 tap filter). So if the first zero is say z1, then the Z transform of the system would be (1-z1*z^-1). Hence, the impulse response of the first filter (say h1[n]) would be the inverse Z tranform of (1-z1*z^-1) which is h1[n] = 1, -z1. Similarly, each zero of the entire system can be modelled as a 2 tap FIR filter. The total transfer function of the system is basically the product of the individual transfer functions i.e. (1-z1*z^-1)(1-z2*z^-1).....((1-z13*z^-1)). So essentially, it is a cascade of several 2 tap filters. The time domain impulse response of the entire system is the convolution of the series of impulse reponses i.e. h[n] = h1[n] * h2[n] * .....h13[n]. where h1[n] = [1, -z1], h2[n] = [1, -z2].....
So this function does an iterative convolution across each impulse response with the lenght of the convolution output increasing by 1 at every stage. SO after convolving 13(number of system zeros) times, we get a 13+1 tap impulse response. I have verified the output with the scipy.signal.zpk2tf function and it matches exactly. In the next commit, I will try to replace the iterative convolution with simple FFTs of the 2 tap impulse reponses and finally multiplying them and then taking IFFT to get the final finaite impulse reponse of the system.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>