Generate TF using FFT and IFFT
In the previous commit, I had generated the final system impulse response by successive convolution of several cascaded 2 tap impulse responses. The process of convolution is compute heavy. Convolution operation can be replaced by multiplication in frequency domain followed by IFFT. FFT/IFFT functions are well optimized and are far less compute intense as compared to convolution operation. Here's how the FFT/IFFT based method is used to generate the final impulse response of the modified zero function.
Let us again assume a 13 zero system whose zeros have been modified. We need to generate an impulse reponse from the zeros. We know that the length of the impulse response is the number of zeros + 1. Now, in the previous commit, I had explained how the individual impulse responses of each of the 13 filters are constructed. Since, the final impulse response will be of length 14 (13 + 1), we perform a 14 point FFT to each of the 13 impulse responses(tap length = 2). I then multiply all the 13 FFTs(each of length 14) to get one single 14 point spectrum. I then take the IFFT of the 14 length spectrum to get the 14 tap impulse response. This method is used to get the impulse response of a system without actual convolution but just using FFTs and IFFTs which are very optimized algorithms. I have checked the output of this method with both the inbuilt function as well as the convolution based approach. The values match very closely upto 6-7 decimals.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>