Replace window function with Vandermonde Conjugate as starting point for zeros
In this commit, I have removed the existing window function as the starting point for the zero manipulation. Instead, I choose the conjugate of the vandermonde matrix super set as a starting point. Towards this I have made the following changes.
Added a function to pre-compute the zeros of the conjugate of the vandermonde matrix for all the quantized digital frequencies from -pi to pi. For example, in this case, the original number of samples is 168 and so for each of the 168 frequency indices, I generate a 14 sample signal and this is called the Vandermonde matrix superset. Since, the response that we wish to generate is like a pseudo inverse solution, it should have the conjugation already applied to it and so we take a conjugate of the vandermonde matrix superset. Also, for ease of indexing, I apply an FFT shift, so that both the positive and negative freqeuncies can be indexed from 0 to 167. Now, we have a matrix of shape 14x168 with both conjugation and FFT shift applied. Now we compute the zeros of each column of this matrix and store them as an LUT. So we have a zeroMatrix of shape 13x168. This will be the initial point of our manipulation with zeros. Now, based on the signal frequency and the interferers, I will sample the zeroMatrix with the index corresponding to the signal frequency. Then, will manipulate the zeros of this system and replace them with the closest interferers. This way, the signal frequency will be intact(since we have chosen the correct column corresponding to the signal freqeuncy) and the zeros and pushed to the interferers. With this change, I have plotted the response of the zero manipulated transfer function and compared it with the vandermonde matrix pseudo inverse response. They match quite closesly upto about 10 bins of separation on a 168 scale. This was not happening with the kaiser window that I was using earlier. This ws because, the kaiser window zeros can be used to manipulate and push the interferers to 0 but it doesn't bother about boosting the signal component. But by choosing the conjugate of the signal freqeuncy as the starting point instead of the window function, we take care of the signal component as well.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>