Compare performance of LeastSquares, DFT, FIR
In this commit, I have written a script to comprehensively compare the performance of Least Squares(LS), DFT, FIR filter (designed by modifying zeros) in estimating the amplitude and phase of the two sinusoidal signals as the separation between the frequencies varies from 0 to Fs or 0 bins to N FFTs bins. Through a montecarlo simulation, we analyze and compare the performance of these 3 methods across different SNR conditions. The results show that, the performance of the FIR filter comes very close to the LS while the DFT performs poorly due to its inherent poor rejection property. I have shown that the Zero modified FIR transfer function performs almost as good as LS. This is a very useful results since we can deploy a simple correlator to estimate the signal amplitide and phase without having to perform a pseudo inverse which is a compute heavy operation. So the action of the Zero Modified FIR tranfer function is similar to that of the LS i.e enhance the signal tone while supressing the interferer tones but with the added benefit of simple compute. The results/plots look promising and this is a break through result.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>