Fix bug in range bin migration modeling term in script
In this commit, I have fixed a bug in modeling the range bin migration
term in the script 'rangeBinMigrationInducedPhaseJumps_radarBaseBandModel.py'. The correct modeling term was already present but it was commnented out. In its place, I had added another term. But this was incorrect. So I have reverted back to the correct range bin migration modeling term. With this change, all is working fine now. I see phase jumps (~pi radians) occuring whenever there is a range bin migration and this is happening for all the bandwidths (500MHz, 1GHz, 4GHz). This measn that all is fine.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>