MIMO coefficient estimation under DDMA scheme
In this commit, I have successfully been able to model the DDMA scheme with simulataneous transmission from 4 Txs each separated by lamda/2 and received by 8 Rxs each separated by 2lamda. I will update the branch with the pdf explainig the mathematical model used to derive the DDMA and estimate the coefficients. For now, I have modelled only a single Doppler in a given range bin so effectively it becomes 4 Dopplers(since 4 Txs). The phase variations across the 4 Dopplers sandwiched between the phase variations across the 8 Rxs helps synthesize a 32 element ULA with a perfect linear phase with slope corresponsing to the programmed 10 degrees object. Currently, I have used 280 ramps in the MIMO segment. Will check later how much this can be reduced to. The challenge I faced in the MIMO coefficient estimation was that when I computed a simple 280 point Doppler FFT(since 280 ramps), the phase of each of the modulated Doppler was not accurate enough to synthesize a perfect liner phase ULA (of 32 elements). However, when I performed a 2048 point Doppler FFT on the 280 Doppler samples, the phase estimate was very accurate and the synthesized ULA was perfectly linear. This is because, in the DDMA scheme, where we rely on the phase variations across the Dopplers, it is very important to have an accurate estimation of the Phase for each of the Doppler peak. When we perform a 280 point Doppler FFT, the bin resolution is poor and hence due to bin quantization, there will be a residual phase error(from the bin quantization) on the Doppler peak. This corrupts the phase estimate. However, when we perfrom a highly oversampled Doppler FFT (2048 point FFT in this case), the accuracy of the Doppler peak improves and subsequently, the accuracy of the Magnitude and phase of the Doppler peak is also very good (due to very minimal bin qunatization error and hence minimal residual phase error).

In the TDMA case, the accuracy of the Doppler correction Matrix (DCM)is very critical in synthesizing a perfect ULA. The accuracy of the DCM is dependent on the accuracy of the Doppler estimate. If the Doppler is not accurate enough, then there will be residual phase on the DCM and this will corrupt the ULA.
For the DDMA case, the accuracy of the Doppler phase is critical in synthesizing a perfect ULA. So when a we take a 280 point Doppler FFT, the Doppler phase accuracy is poor as compared to when we take a 2048 point Doppler FFT.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>