Added Doppler to the DDMA simulation
In the previous commit, I had not modelled Doppler in the DDMA simulation. So in essence, it was just a 0 Doppler component getting modulated with the Phase codes of each Tx. In this commit, I have modelled the Doppler as well. It doesn't make much difference whether there is Doppler or not since anyways, even if we have a static target, we get the modulated components in the Doppler Spectrum. But for modelling completeness and also to extend later onto multiple Dopplers in a given range bin, having a Doppler will be useful and hence I have incorporated Doppler also. So when synthesizing the ULA by sampling the Doppler indices, I add the true Doppler index(Dopp/DoppRes) to the index obtained from the phase change per Tx that is programmed. These new Doppler indices have to be modulo 2048(since we perform a 2048 point Doppler FFT) and are used to sample the phase values from the DOppler spectrum.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>