Give directions to verify ABF scenario
IN this commit, I have given some steps (in the form of comments) to
check if the DDMA MIMO has been modelled correctly. To check this, we
test if the DDMA is a special case of ABF. By following the instructions
in the comments, we can verify that the DDMA modelling has indedded been
done quite precisely and the ABF is infact a degenerate case of the MIMO
DDMA. I have successfully verified this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>