Model 2 Dopplers in a given range for the DDMA scheme
In this commit, I have another script which models 2 Dopplers in a given
range. It is just an extension of the single Doppler script using
superposition for 2 Dopplers. The results are on expected lines and I'm
able to estimate the Azimuth angle for both the Dopplers from the DDMA
scheme. However, in this case, the 2 Dopplers(and each of their 4
modulated components) do not come too close to each other and do not
overlap. So there is no corruption in the phases for the ULA. But if the
DOpplers start to come close to each other or they start to overlap,
then the Azimuth ULA phases might be corrputed. In the subsequent
commits, I will try to highlight this issue.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>