Ammend the observation with respect to DNL floor raise
In this commit, I have ammended/corrected my previosu observation that
with very Tx added to the DDMA scheme, the DNL noise floor increases by
3 dB. This observation is wrong. The correct observation is that the DNL
noise floor raises by 6dB for every doubling of Txs in the simultaneous
DDMA scheme. This is because both the signal and noise add in voltage
and hence we obatin a 6dB raise for every doubling of Txs. Hence the DNL
nosie floor also raises by 6dB when we move from 2 Tx DDMA to 4 Tx DDMA.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>