Incorporated the Range Bin Migration term to the DDMA model
In this commit, I have incorporated the range bin migration term to the DDMA model. Due to the presence of the range bin migration term, I now slice the chirp samples using the updated range bins (based on true velocity) for each chirp. Post this, I apply the phase jump correction factor(PI-PI/N) to the chirp samples wherever there is a bin migration. Post this PI phase correction, the Doppler phase now becomes linear. However, I noticed something interesting when plotting this phase against the Tru Doppler phase. I noticed that even though the slop of the phase becomes linear post the PI phase correction, the slope doesnt match the true ground truth Doppler phase. In toehr words, if we were to plot the Doppler spectrum post the PI phase correction, the SLLs are good however, the location of the peak doesn't match the true ground truth Doppler spectrum. There is still an uncorrected modulation term. Going through one of the range bin migration papers, I was able to figure out this modulation term. Below is a small recap.
When there is range bin migration, we have shown that there will be a phase jump of PI-PI/N which needs to be corrected to get a linear phase. But there is a little more to that. Post the PI-PI/N phase correction, the Doppler phase or ULA phase becomes linear but the slope still doesn't match the true ground truth Doppler slope. In other words, the Doppler frequency estimated post the PI-PI/N phase correction will still be off from the true Doppler frequency(by a few Hz). This delta in Doppler frequency is further exaggerated with larger chirp sweep bandwidths i.e the delta in Doppler frequency (between ground truth and estimated) will be larger for higher chirp bandwidths. So, to take care of this, we need to apply another demodulation term post the PI-PI/N phase correction. With this phase correction, the delta becomes much smaller. This demodulation term is given by exp(1j*2PI*(Bv/c)*n*Tc), where 'B' is the chirp bandwidth, 'v' is the true unaliased velocity of the target, 'c' is the speed of light, 'n' is the chirp index and 'Tc' is the inter-chirp time. In this commit, I have incorporated this demodulation term and now the Doppler peak is exactly matching the ground truth peak. This term can be derived by using a slightly different mathematical model of the RADAR equation and this model is available in a paper that I will be attaching later on. I will also derive the same model from my end.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>