Not correct for Doppler modulation on the chirp samples but rather correct/update the Doppler bin to sample
In this commit, instead of correcting for the Doppler modulation due to range movement on the chirp samples, I'm now updating the doppler bin to sample by accounting for the Bv/c term. Both are correct. Either we remove the modulation on the signal and sample the Doppler + its modualtion bins or we don't remove the Doppler modulation on the chirp samples but update the doppler bin to sample by accounting for the Bv/c term. In the single Doppler DDMA script, I'm doing the former and in this generalized script I'm doing the latter.

Some of the other changes made in this commit are as follows:
1. Randomly selecting the number of Dopplers per range from anywhere between 1 to 4
2. Printing the DDMA and Doppler FFT gain so that it will help in choosing the target range snr appropriately.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>